undoing last attempted fix

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -17,6 +17,6 @@ int main(void)
 			ret = i;
 		}
 	}
-	printf("%ld\n", (int)ret);
+	printf("%ld\n", ret);
 	return (0);
 }
